Resolve error

See merge request s02-ai-sub2/s02p22a405!81

Former-commit-id: 60ce3768e257034c31ed9f4a016517b7f2f443e1

+++ data/preprocess.py
@@ -52,25 +52,34 @@ def get_data_file():
 
 
 def save_tokenizer(data_path, caption_num_words=5000):
-    data = np.load(data_path)
-    captions = data[:, 2:]
+    tokenizer_path = os.path.join(BASE_DIR, 'tokenizer.pkl')
+    if not os.path.exists(tokenizer_path):
+        data = np.load(data_path)
+        captions = data[:, 2:]
 
-    captions = np.squeeze(captions, axis=1)
-    captions = ['<start>' + cap + ' <end>' for cap in captions]
+        captions = np.squeeze(captions, axis=1)
+        captions = ['<start>' + cap + ' <end>' for cap in captions]
 
-    tokenizer = tf.keras.preprocessing.text.Tokenizer(num_words=caption_num_words + 1,
-                                                      oov_token='<unk>',
-                                                      lower=True,
-                                                      split=' ',
-                                                      filters='!"#$%&()*+.,-/:;=?@[\]^_`{|}~ ')
+        tokenizer = tf.keras.preprocessing.text.Tokenizer(num_words=caption_num_words + 1,
+                                                          oov_token='<unk>',
+                                                          lower=True,
+                                                          split=' ',
+                                                          filters='!"#$%&()*+.,-/:;=?@[\]^_`{|}~ ')
 
-    tokenizer.fit_on_texts(captions)
-    tokenizer.word_index['<pad>'] = 0
-    tokenizer.index_word[0] = '<pad>'
+        tokenizer.fit_on_texts(captions)
+        tokenizer.word_index['<pad>'] = 0
+        tokenizer.index_word[0] = '<pad>'
 
-    with open('./datasets/tokenizer.pkl', 'wb') as f:
-        pickle.dump(tokenizer, f, protocol=pickle.HIGHEST_PROTOCOL)
-        
+        with open('./datasets/tokenizer.pkl', 'wb') as f:
+            pickle.dump(tokenizer, f, protocol=pickle.HIGHEST_PROTOCOL)
+    else:
+        with open(tokenizer_path, 'rb') as f:
+            tokenizer = pickle.load(f)
+
+    return tokenizer
+
+    
+# Find the maximum length of any caption in our dataset
 def calc_max_length(tensor):
     return max(len(t) for t in tensor)
 
@@ -81,11 +90,6 @@ def change_text_to_token(train_captions):
     max_length = calc_max_length(train_seqs)
     cap_vector = tf.keras.preprocessing.sequence.pad_sequences(
         train_seqs, padding='post')
-    print()
-    print('전처리 step 2')
-    print('캡션 텍스트 토큰화 ex) ')
-    print('1: ', cap_vector[:1])
-    print('2: ', cap_vector[1:2])
     return cap_vector, max_length
 
 
@@ -107,7 +111,9 @@ def get_image_datasets(img_name_vector):
 
 
 def map_func(img_name, cap):
-    img_tensor = np.load(img_name.decode('utf-8') + '.npy')
+    feature_name = os.path.basename(img_name).decode(
+        'utf-8').replace('jpg', 'npy')
+    img_tensor = np.load((os.path.join(BASE_DIR, 'features', feature_name)))
     return img_tensor, cap
